+++ src/lib.rs
@@ -7,15 +7,15 @@ mod interpreter;
 pub mod commands;
 
 pub struct Rolls {
-    pub max: u32,
-    pub min: u32,
-    pub dice: Vec<u32>,
+    pub max: u64,
+    pub min: u64,
+    pub dice: Vec<u64>,
 }
 
-pub fn roll_dice(count: u32, sides: u32) -> Rolls {
-    let mut dice: Vec<u32> = Vec::new();
-    let mut min = u32::MAX;
-    let mut max = u32::MIN;
+pub fn roll_dice(count: u64, sides: u64) -> Rolls {
+    let mut dice: Vec<u64> = Vec::new();
+    let mut min = u64::MAX;
+    let mut max = u64::MIN;
 
     for _ in 0..count {
         let die = rand::thread_rng().gen_range(1..=sides);
